Also consider `rdf_[un]linked_shared_data` in `update_wikidata` (#197)
So far, the command `update-wikidata` only considered the parts `rdf_added_data` (triples that must be added) and `rdf_deleted_data` (triples that must be removed) of each message from the Wikidata update stream.

Now, the command also consider `rdf_linked_shared_data` (triples that must be added only in case they have not already been added, which might be the case) and `rdf_unlinked_shared_data` (triples that should be removed, but do not harm if they are not removed). See https://wikitech.wikimedia.org/wiki/Wikidata_Query_Service/Streaming_Updater/Public_Update_Stream

NOTE: With this change, messages where the value for `"operation"` is one of `"import"`, `"diff"` or `"reconcile"` are now processed properly. However, messages with `"operation": "delete"` are still not processed properly (because for these, all four of the `rdf_...` above are empty and a `DELETE WHERE` query would be needed).