Consider `delete` operations in Wikidata update stream (#206)
So far, the `delete` operations were simply ignored. Now they are considered and executed in one `DELETE { ... } WHERE { ... }` at the end of each batch. When an added triple is encountered that has a subject that was deleted earlier in the batch, the batch is  finished and that added triple is processed in the next batch. It turns out that this happens rarely, even for a batch size of `100,000`.

On the side, improve the code structure, the logging, and the handling of Ctrl+C.